--- a/python-sdk/tests/function_executor/utils.py
+++ b/python-sdk/tests/function_executor/utils.py
@@ -20,6 +20,7 @@ class FunctionExecutorServerTestCase(unittest.TestCase):
             [
                 "indexify-cli",
                 "function-executor",
+                "--dev",
                 "--function-executor-server-address",
                 cls.FUNCTION_EXECUTOR_SERVER_ADDRESS,
                 "--indexify-server-address",